Kerberos.Net.Entities.Krb.KrbPrincipalName: Generate CNames that match MIT Kerberos kinit
At present, Kerberos.NET always sends AS_REQ requests to the KDC with a CName
consisting of the full principal name with type NT_ENTERPRISE.  This breaks
when authenticating to an MIT Kerberos KDC, especially when service principals
are used.

This commit makes Kerberos.NET process such names as NT_PRINCIPAL principals
with either one or two name parts, depending on whether the principal contains
a slash in the name or not.  This means that the system authenticates correctly.